// client/Game/src/GameFrame/Menu/MenuLayerType.ts
// ===========================
// 层级类型
// ---------------------------
export enum MenuLayerType
{
    // 主界面
    Home = 1,

    // 主界面 上一层
    HomeTop = 2,

    // 模块
    Module = 3,

    // 模块
    ModuleTop = 4,

    // 主UI
    MainUI = 5,

    // 对话框
    Dialog = 6,

    // 引导
    Guide = 7,

    // 加载面板
    Loader = 8,

    // GM
    GM = 9,

    // 浮动对话框
    FloatMsg = 10,
}